app/requirements.txt
streamlit>=1.32.0
python-docx>=1.1.0
PyPDF2>=3.0.0
trafilatura>=1.8.0
newspaper3k>=0.2.8
gnews>=0.3.7
rapidfuzz>=3.6.0
openai>=1.12.0
pyyaml>=6.0
requests>=2.31.0
